requirements.txt
# Core ADK Dependencies
google-adk>=0.3.0
google-generativeai>=0.8.0

# LLM & AI
anthropic>=0.40.0  # For potential Claude integration
litellm>=1.50.0    # Multi-model support

# Web Framework
streamlit>=1.39.0
streamlit-option-menu>=0.3.13

# Data Processing
pandas>=2.2.0
openpyxl>=3.1.5    # Excel export
xlsxwriter>=3.2.0

# Database
sqlalchemy>=2.0.35

# Web Scraping (for real data phase)
beautifulsoup4>=4.12.3
requests>=2.32.0
selenium>=4.26.0
lxml>=5.3.0

# Utilities
python-dotenv>=1.0.1
pydantic>=2.9.0
pyyaml>=6.0.2

# Logging & Monitoring
structlog>=24.4.0
python-json-logger>=2.0.7

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0

# Development
black>=24.10.0
flake8>=7.1.0
mypy>=1.13.0

# Deployment
gunicorn>=23.0.0
